Squashed commit of the following:
commit cfe378043cb76c929db519b82c7327ae007c638a
Author: Andre McGrail <andrem@unity3d.com>
Date:   Thu Oct 20 16:22:16 2022 +0200

    Updated package json

commit c0133fa599f2ce3cac70daaddff7df8f16e45728
Author: Andre McGrail <andrem@unity3d.com>
Date:   Thu Oct 20 15:11:01 2022 +0200

    start of water settings

commit 756aa173f5fab385a9bc7f900919a63357bb05c3
Author: Andre McGrail <andrem@unity3d.com>
Date:   Mon Jul 11 13:33:52 2022 +0200

    keyword tweaks

commit d8525caec60619ca18a07147769194da7c247bab
Author: Andre McGrail <andrem@unity3d.com>
Date:   Wed Jun 15 10:07:33 2022 +0200

    Tweaks

    - added scrolling to buoyancy debug window
    - added a water event script to try provide a way for users to query a point in space and it's interaction with the water
    - tweaked cpu water code to be more async

commit 84d23d57148adf0a637a610c52adfeccedc76fe2
Author: Andre McGrail <andrem@unity3d.com>
Date:   Tue May 3 15:21:04 2022 +0200

    Final tweaks for initial usage

+++ Runtime/Shaders/GerstnerWaves.hlsl
@@ -85,8 +85,6 @@ inline void SampleWaves(float3 position, half opacity, out WaveStruct waveOut)
 {
 	waveOut = (WaveStruct)0;
 	half2 pos = position.xz;
-	//waveOut.position = 0;
-	//waveOut.normal = 0;
 	half waveCountMulti = 1.0 / _WaveCount;
 	opacity = saturate(opacity);
 
@@ -116,7 +114,7 @@ inline void SampleWaves(float3 position, half opacity, out WaveStruct waveOut)
 		waveOut.foam += wave.foam;
 	}
 	waveOut.position *= opacity;// opacityMask;
-	waveOut.normal *= half3(opacity, 1, opacity);
+	waveOut.normal *= float3(opacity, 1, opacity);
 	waveOut.foam *= waveCountMulti * opacity;
 }
